move write var handling to state module
